*** empty log message ***
git-svn-id: http://svn.jboss.org/repos/jbosstools/trunk@885 a97e2381-89e5-4abb-bab3-167db6db766c

+++ cache/plugins/org.jboss.ide.eclipse.jbosscache/src/org/jboss/ide/eclipse/jbosscache/actions/DeleteConfigurationAction.java
@@ -6,8 +6,20 @@
  */
 package org.jboss.ide.eclipse.jbosscache.actions;
 
+import java.io.File;
+
 import org.eclipse.jface.dialogs.IDialogConstants;
 import org.eclipse.jface.dialogs.MessageDialog;
+import org.eclipse.jface.viewers.ISelection;
+import org.eclipse.swt.SWT;
+import org.eclipse.swt.events.SelectionEvent;
+import org.eclipse.swt.events.SelectionListener;
+import org.eclipse.swt.graphics.Image;
+import org.eclipse.swt.layout.GridData;
+import org.eclipse.swt.widgets.Button;
+import org.eclipse.swt.widgets.Composite;
+import org.eclipse.swt.widgets.Control;
+import org.eclipse.swt.widgets.Shell;
 import org.jboss.ide.eclipse.jbosscache.ICacheConstants;
 import org.jboss.ide.eclipse.jbosscache.model.cache.ICacheRootInstance;
 import org.jboss.ide.eclipse.jbosscache.model.internal.TreeCacheManager;
@@ -30,29 +42,100 @@ public class DeleteConfigurationAction extends AbstractCacheAction
    {
       super(view, actionId);
    }
+   
+   private class CustomMessageDialog extends MessageDialog{
+
+      private boolean isCheck = false;
+      private ICacheRootInstance rootInstance;
+      
+      public CustomMessageDialog(Shell parentShell, String dialogTitle, Image dialogTitleImage, String dialogMessage, int dialogImageType, String[] dialogButtonLabels, int defaultIndex)
+      {
+         super(parentShell, dialogTitle, dialogTitleImage, dialogMessage, dialogImageType, dialogButtonLabels, defaultIndex);
+      }
+
+      protected Control createCustomArea(Composite parent)
+      {
+         Button btn = new Button(parent,SWT.CHECK);
+         btn.setText("Configuration file will be removed from the file system");
+         btn.setLayoutData(new GridData(GridData.FILL_HORIZONTAL));
+         
+         if(rootInstance.isRemoteCache())
+            btn.setEnabled(false);
+         else
+            btn.setEnabled(true);
+         
+         btn.addSelectionListener(new SelectionListener(){
+
+            public void widgetSelected(SelectionEvent e)
+            {
+               Button btn = (Button)e.widget;
+               if(btn.getSelection())
+                  isCheck = true;
+               else
+                  isCheck = false;
+               
+            }
+
+            public void widgetDefaultSelected(SelectionEvent e)
+            {
+               widgetSelected(e);
+               
+            }
+
+            
+         });
+           
+         return btn;
+      }
+      
+      public boolean isCheck(){
+         return isCheck;
+      }
+      
+      public void setRootInstance(ICacheRootInstance rootInstance){
+         this.rootInstance = rootInstance;
+      }
+            
+   }
 
    /**
     * Delete configuration action
     */
    public void run()
    {
-      MessageDialog msgDialog = new MessageDialog(getTreeViewer().getShell(), CacheUtil
+      ICacheRootInstance rootInstance = (ICacheRootInstance) getTreeViewer().getSelection();
+      
+      CustomMessageDialog msgDialog = new CustomMessageDialog(getTreeViewer().getShell(), CacheUtil
             .getResourceBundleValue(ICacheConstants.TREECACHEVIEW_DELETE_CONFIGURATION_DIALOG_TITLE), null, CacheUtil
             .getResourceBundleValue(ICacheConstants.TREECACHEVIEW_DELETE_CONFIGURATION_DIALOG_MESSAGE),
             MessageDialog.QUESTION, new String[]
             {IDialogConstants.YES_LABEL, IDialogConstants.NO_LABEL}, 1);
-
+      
+      msgDialog.setRootInstance(rootInstance);
+     
       int delete = msgDialog.open();
+      
 
       if (delete == IDialogConstants.OK_ID)
-      {
-         ICacheRootInstance rootInstance = (ICacheRootInstance) getTreeViewer().getSelection();
-         if (rootInstance.isConnected())
+      {  
+         if(!rootInstance.isRemoteCache())
          {
-            TreeCacheManager manager = rootInstance.getTreeCacheManager();
-            TreeCacheManager.stopService_(manager);
-            TreeCacheManager.destroyService_(manager);
+            if (rootInstance.isConnected())
+            {
+               TreeCacheManager manager = rootInstance.getTreeCacheManager();
+               TreeCacheManager.stopService_(manager);
+               TreeCacheManager.destroyService_(manager);
+            }
+            
+            boolean isCheck = msgDialog.isCheck();
+            if(isCheck)
+            {
+               File file = new File(rootInstance.getRootConfigurationFileName());
+               if(file.exists())
+                  file.delete();
+            }
          }
+                  
          rootInstance.deleteRoot();
       }
    }
